requirements.txt
# Core dependencies for SAP S/4HANA Data Quality Workflow
# This file contains dependencies for local development and testing

# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Data Processing
pandas==2.2.0
numpy==1.26.4
pyyaml==6.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.27

# Monitoring & Logging
prometheus-client==0.20.0
structlog==24.1.0

# Utilities
python-dotenv==1.0.1
click==8.1.7
requests==2.31.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Development tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0 